+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -106,6 +106,11 @@ New checks
 
   FIXME: add release notes.
 
+- New :doc:`hooks-entry-points
+  <clang-tidy/checks/hooks-entry-points>` check.
+
+  FIXME: add release notes.
+
 - New :doc:`hooks-guard-called
   <clang-tidy/checks/hooks-guard-called>` check.